chore: move layouts to dedicated folder
