+++ .vitepress/theme/layouts/Home.tsx
@@ -1,5 +1,5 @@
 import {defineComponent} from "vue";
-import {data as posts} from './posts.data';
+import {data as posts} from '../posts.data';
 
 export default defineComponent({
     setup() {
